Log more details on connection failures
The `ConnectionFailedError` is effectively a wrapper around several
kinds of underlying errors, and we'd like to get more details logged
about the initial cause of such errors.

This changes the generation of the `ConnectionFailedError` to include
initial exception details when relevant, and changes the capturing code
in the producer to also log some of those details.