Rescue Errno::ETIMEDOUT to ConnectionFailure
I believe this is the intended behavior of the code: to handle a timeout
as a connection failure which will bubble up into the Producer's retry
semantics.

I believe the code didn't realize that IO.select may return true but the
actual Socker#write may still fail with a timeout, so it wasn't being
handled. This meant it would bubble out as a full exception without any
retry handling. We see this at a low rate on 0.8, but a higher rate on
0.10.

This change makes the smaller fix of adding this exception to the rescue
list, instead of the potential larger fix of removing the select and
manual TimeoutException entirely and just relying on this exception to
indicate a timeout. Making that larger fix would require leaning on the
above assumptions more than I'm comfortable with.